Proyecto_Butterfly_Image_Classification/requirements.txt
Flask
tensorflow
Pillow
numpy